refactor(admin): split health route modules
Why:
- keep the Quantercise health import path stable while separating data loading, skeleton UI, and health panels
- reduce a mixed server/rendering file and isolate failed-event row rendering from fetch orchestration

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/health/page.tsx apps/admin/src/app/quantercise/health/health-sections.tsx apps/admin/src/app/quantercise/health/health-content.tsx apps/admin/src/app/quantercise/health/health-panels.tsx apps/admin/src/app/quantercise/health/health-skeleton.tsx apps/admin/src/app/quantercise/health/failed-event-actions.tsx
- git diff --check -- apps/admin/src/app/quantercise/health/page.tsx apps/admin/src/app/quantercise/health/health-sections.tsx apps/admin/src/app/quantercise/health/health-content.tsx apps/admin/src/app/quantercise/health/health-panels.tsx apps/admin/src/app/quantercise/health/health-skeleton.tsx apps/admin/src/app/quantercise/health/failed-event-actions.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
